# docs(perf): use `swc-loader`