--- a/website/package.json
+++ b/website/package.json
@@ -7,14 +7,18 @@
     "docusaurus": "docusaurus",
     "serve": "docusaurus serve"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "@types/node": "^22.5.0"
+  },
   "dependencies": {
     "@docusaurus/core": "3.4.0",
     "@docusaurus/plugin-client-redirects": "3.4.0",
     "@docusaurus/preset-classic": "3.4.0",
     "@easyops-cn/docusaurus-search-local": "0.44.4",
+    "@swc/core": "^1.7.14",
     "clsx": "^2.0.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "swc-loader": "^0.2.6"
   }
 }