## Summary

- migrate [configuration file](https://docusaurus.io/docs/configuration#syntax-to-declare-docusaurus-config) to typescript to get editor support
- use [swc-loader](https://github.com/swc-project/pkgs/tree/main/packages/swc-loader), a drop-in webpack babel-loader replacement to improve build speed by ~16%.

## Description

| method            | babel-loader (default) | swc-loader (new)         |
| ----------------- | ---------------------- | ------------------------ |
| with minification | 46.22s                 | 38.59s (83% of original) |
| no minification   | 42.60s                 | 34.73s (81% of original) |

<details><summary>Log</summary>

```sh
$ yarn docusaurus clear
$ time yarn build --no-minify
   
[INFO] Use `npm run serve` command to test your build locally.
Done in 42.32s.

________________________________________________________
Executed in   42.60 secs    fish           external
   usr time   59.48 secs  861.00 micros   59.48 secs
   sys time    6.54 secs  143.00 micros    6.54 secs

$ yarn docusaurus clear
$ time yarn build
Done in 45.91s.

________________________________________________________
Executed in   46.22 secs    fish           external
   usr time   99.91 secs    0.00 micros   99.91 secs
   sys time    9.69 secs  965.00 micros    9.69 secs

$ yarn docusaurus clear
$ time yarn build --no-minify
Done in 34.43s.

________________________________________________________
Executed in   34.73 secs    fish           external
   usr time   50.03 secs    0.04 millis   50.03 secs
   sys time    4.80 secs    1.17 millis    4.79 secs

$ yarn docusaurus clear
$ time yarn build
Done in 38.30s.

________________________________________________________
Executed in   38.59 secs    fish           external
   usr time   89.68 secs    0.00 micros   89.68 secs
   sys time    7.01 secs  870.00 micros    7.01 secs
```


</details> 

## Additional context

guides followed: 
- https://johnnyreilly.com/faster-docusaurus-build-swc-loader
- https://github.com/facebook/docusaurus/issues/4765#issuecomment-841135926